Remove command pattern usages
